+++ CHANGELOG.md
@@ -2,13 +2,14 @@
 
 ENHANCEMENTS:
 - `msgraph` resources and data sources now support `retry` configuration to handle transient failures.
+- `msgraph` resource and data source: support for `timeouts` configuration block
 
 FEATURES:
 - **New Resource**: msgraph_update_resource
 
 BUG FIXES:
 - Fixed an issue where `msgraph_resource` resource did not wait for the resource to be fully provisioned before completing.
-
+- Fixed an issue with the `msgraph_resource` resource could not detect resource drift.
 
 ## 0.1.0